Corrijo Time Stamp

+++ ProyectoFinalElectricidadSeret/Models/Accountancy/BancContac.cs
@@ -7,8 +7,9 @@ namespace ProyectoFinalElectricidadSeret.Models
 {
     public partial class BancContac
     {
-        public decimal BycCodbce { get; set; }
+        public int BycCodbce { get; set; }
         public decimal BycCodcon { get; set; }
+
         public virtual Banco BycCodbceNavigation { get; set; }
         public virtual Contacto BycCodconNavigation { get; set; }
     }
